requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
MarkupSafe==2.1.4
mongoengine==0.27.0
motor==3.3.2
orjson==3.9.12
passlib==1.7.4
pycparser==2.21
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
PyJWT==2.8.0
pymongo==4.6.1
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.35.1
tortoise-orm==0.20.0
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
